Update README.md
When the cells in the body start growing beyond control, it takes a form of the disease known as cancer. When this disease occurs in the lungs, then it is termed as lung cancer. The cancer spreads all over the lymph nodes, and in some instances, it might apply to various parts of the body. It is also possible that cancer can reach your lungs from other organs. And when the disease starts spreading to multiple organs, it is called metastases.